Small program to create a "Work Buckets" column in trello
It walks every card in New, Backlog, Next, In Progress, and Pending
Upstream looking for [*].  If it finds * in the 'Work Buckets' column it
will add the card to the checklist for that card.

Cards with no buckets will get listed on a card called 'None' (if it
exists)

If there is no card called * in the Work Buckets column but there is a
card called 'None' * will be added as an Unknown Bucket to the None
card.